+++ server.js
@@ -17,7 +17,8 @@ var serve = require("./lib/app"),
 module.exports = function(opts, callback) {
   var app = express().disable("x-powered-by"),
       tilelive = require("tilelive-cache")(require("tilelive"), {
-        size: process.env.CACHE_SIZE || opts.cacheSize
+        size: process.env.CACHE_SIZE || opts.cacheSize,
+        sources: process.env.SOURCE_CACHE_SIZE || opts.sourceCacheSize
       });
 
   callback = callback || function() {};